requirements.txt
# Auto-generated file
pyyaml
pandas 
numpy
tqdm
torch
torchvision
ogb
torch_geometric
datasets
transformers
matplotlib
statsmodel